* generic/tclCkalloc.c: fix warnings from gcc build configured with 	* generic/tclCompile.c: --enable-64bit --enable-symbols=all. 	* generic/tclExecute.c:

+++ generic/tclCkalloc.c
@@ -213,7 +213,7 @@ ValidateMemory(
 				 * been printed */
 {
     unsigned char *hiPtr;
-    int idx;
+    size_t idx;
     int guard_failed = FALSE;
     int byte;
 
@@ -223,7 +223,7 @@ ValidateMemory(
 	    guard_failed = TRUE;
 	    fflush(stdout);
 	    byte &= 0xff;
-	    fprintf(stderr, "low guard byte %d is 0x%x  \t%c\n", idx, byte,
+	    fprintf(stderr, "low guard byte %d is 0x%x  \t%c\n", (int)idx, byte,
 		    (isprint(UCHAR(byte)) ? byte : ' ')); /* INTL: bytes */
 	}
     }
@@ -244,7 +244,7 @@ ValidateMemory(
 	    guard_failed = TRUE;
 	    fflush(stdout);
 	    byte &= 0xff;
-	    fprintf(stderr, "hi guard byte %d is 0x%x  \t%c\n", idx, byte,
+	    fprintf(stderr, "hi guard byte %d is 0x%x  \t%c\n", (int)idx, byte,
 		    (isprint(UCHAR(byte)) ? byte : ' ')); /* INTL: bytes */
 	}
     }
